refactor: HTTP client configuration and error handling utilities
